Oops - take a copy of ath_tx_status from the buffer before the TX processing is done.
The aggregate path was definitely accessing 'ts' before it was actually
being assigned.

This had the side effect of over-filtering frames, since occasionally that
bit would be '1'.

Whilst here, do the same thing in the non-aggregate completion function -
as calling the filter function may also invalidate bf.

Pointy hat to: adrian, for not noticing this over many, many code reviews.


git-svn-id: svn://svn.freebsd.org/base/head@240677 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f