zfsboot: simplify probe_drive() a little bit
The first discovered pool, whether it covers the whole boot disk or not,
is going to be first in zfs_pools list.  So there is no need at all
for spapp parameter.

This commit also fixes a bug where NULL would be assigned to NULL
pointer when probe_drive was called  with the spapp parameter of NULL.

MFC after:	21 days


git-svn-id: svn://svn.freebsd.org/base/head@241294 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f