Since it's not immediately obvious whether the current TX path handles fragment rate lookups correctly, add a comment describing exactly that.
The assumption in the fragment duration code is the duration of the next
fragment will match the rate used by the current fragment.  But I think
a rate lookup is being done for _each_ fragment.  For older pre-sample
rate control this would almost always be the case, but for sample
it may be incorrect more often then correct.


git-svn-id: svn://svn.freebsd.org/base/head@242144 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f